voice agent/day 10/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
assemblyai
python-dotenv
